Extract the marquee drag mechanics into a shared engine
